syk: Allow the user to specify gamma matrix normalization
Allow the user to specify whether he wants {gamma_i,gamma_j}=delta_ij
or {gamma_i,gamma_j} = 2delta_ij. The latter is convenient when
you want to avoid normalizing the state vector after applying the
gamma_i operator to it. The default is the standard normalization
(the former).