cpuimpl: Implement exact diagonalization using Eigen3
Provide an alternative state evolution algorithm using by exactly
diagonalizing the Hamiltonian first and exponentiate its eigenvalues.
This is implemented for the CPU backend at this point in time. We
also provide an interface with which the end user can retrieve the
eigensystem as well as the matrix exponentials computed during exact
diagonalization.