matexp: Implement caching of results
We implement caching of the matexp results to speed up certain
calculations (such as the Lindblad example). This is implemented
for both CPU and GPU (and they share the same code for handling
the caching). We set a maximum cache size and implement the LRU
eviction policy so to not run out of memory, and we release the
reserved cache space when clearing the cache.

The default cache size is 16GB for CPU and 50% of available VRAM
(at program startup time) for GPU. The former has no practical
meaning because Linux (by default) has memory overcommit enabled.
We provide interfaces for the end user to adjust the cache size.