Improve the ergonomics of get_matrix and matexp
We define SubspaceView to keep track of the spin chain lengths.
This way we don't have to carry around the spin chain lengths
whenever we call get_matrix and matexp. Use variadic templates
to pass their parameters through parity block forms.

Also add a funclet in the test suite that computes the sparsity
of the dense SYK Hamiltonian matrix. The conclusion is that for
N all the way to 32 (the breakeven point where the number of
terms in the Hamiltonian starts to become less than the dimension
of the Hilbert space) the sparsity never goes to 99% (at best it
is 96%) so we should not convert the dense Hamiltonian matrix to
sparse form.