Add exact diagonalization algorithm for computing the ground state
We move the ground state algorithm to EvolutionAlgorithm::Krylov
and allow user to choose whether to use exact diagonalization or
Krylov to compute the ground state of a Hamiltonian.

The default tolerance of the Krylov ground state algorithm has
been changed to ten times that of the machine precision. We find
that this is adequate and converges sooner.